remove old datepicker
